Refactor material module
Putting nodes in nice places is handled by the new Block abstraction
so the material functions can concentrate on the logic. There are a
bunch of diagrams which illustrate what's happening now too.

There are still some rough spots. Frames in particular are kinda
annoying since its hard to tell where they are and what their sizes
are.
