+++ addons/io_scene_gltf/material/groups.json
